Fixed adcs 1 and 2, now running on update event from PWM interrupt

+++ boards/INV/Inc/App/controls/App_AdaptionGain.h
@@ -4,4 +4,4 @@
 
 struct ControllerValues adaptionGain(
     const struct ControllerValues *const values,
-    const double                         is_ref);
+    const float                          is_ref);
