+++ boards/INV/Inc/App/controls/App_ClarkeParkTransform.h
@@ -3,4 +3,4 @@
 #include "controls/App_ControlLoop.h"
 
 struct DqsValues
-    clarkeParkTransform(struct PhaseValues *phase_currents, double theta);
+    clarkeParkTransform(struct PhaseValues *phase_currents, float theta);
