+++ boards/INV/Inc/App/controls/App_CurrentController.h
@@ -3,14 +3,14 @@
 #include "controls/App_ControlLoop.h"
 
 // DqsValues calculateDqsVoltages(DqsValues * dqs_ref_currents, DqsValues *
-// dqs_currents, double omega, 								double bus_voltage,
+// dqs_currents, float omega, 								float bus_voltage,
 // ControllerValues
 // *
 // id_controller, ControllerValues * iq_controller);
 struct DqsValues calculateDqsVoltages(
     const struct DqsValues *const dqs_ref_currents,
     const struct DqsValues *const dqs_currents,
-    const double                  omega,
-    const double                  bus_voltage,
+    const float                   omega,
+    const float                   bus_voltage,
     struct ControllerValues *     id_controller,
     struct ControllerValues *     iq_controller);
