+++ boards/INV/Inc/App/controls/App_CurrentGeneration.h
@@ -4,6 +4,6 @@
 
 struct DqsValues generateRefCurrents(
     const struct DqsValues *const dqs_ref_currents,
-    double                        omega,
-    double                        vdc_sensor_val,
+    float                         omega,
+    float                         vdc_sensor_val,
     bool *                        fw_flag);
