+++ boards/INV/Inc/App/controls/App_ParkClarkeTransform.h
@@ -4,4 +4,4 @@
 
 struct PhaseValues parkClarkeTransform(
     const struct DqsValues *const dqs_voltages,
-    double                        theta);
+    float                         theta);
