+++ boards/INV/Inc/App/controls/App_TorqueControl.h
@@ -2,11 +2,11 @@
 
 #include "controls/App_ControlLoop.h"
 
-// double torqueControl(double omega_ref, double omega, double torque_ref,
+// float torqueControl(float omega_ref, float omega, float torque_ref,
 // ControllerValues * speed_controller, bool * prev_fw_flag);
-double torqueControl(
-    double                   omega_ref,
-    double                   omega,
-    double                   torque_ref,
+float torqueControl(
+    float                    omega_ref,
+    float                    omega,
+    float                    torque_ref,
     struct ControllerValues *speed_controller,
     const bool               prev_fw_flag);
