+++ boards/INV/Inc/Io/Io_AdcDac.h
@@ -8,10 +8,10 @@ void Io_AdcDac_AdcStart(void);
 void Io_AdcDac_AdcStop(void);
 
 void Io_AdcDac_DacStart(void);
-void Io_AdcDac_DacSetCurrent(double current);
+void Io_AdcDac_DacSetCurrent(float current);
 
-void   Io_AdcDac_GetPhaseCurrents(struct PhaseValues *const phase_currents);
-double Io_AdcDac_GetGpioVal(void);
-double Io_AdcDac_GetBusVoltage(void);
-double Io_AdcDac_GetMotorTemp(void);
-double Io_AdcDac_GetPowerstageTemp(void);
+void  Io_AdcDac_GetPhaseCurrents(struct PhaseValues *const phase_currents);
+float Io_AdcDac_GetGpioVal(void);
+float Io_AdcDac_GetBusVoltage(void);
+float Io_AdcDac_GetMotorTemp(void);
+float Io_AdcDac_GetPowerstageTemp(void);
